Changed 'int sales' to 'long sales'

+++ src/main/java/ru/netology/hw6/stats/StatsService.java
@@ -2,64 +2,64 @@ package ru.netology.hw6.stats;
 
 public class StatsService {
 
-    public int sumSales(int[] sales) {
+    public int sumSales(long[] sales) {
         int sum = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            sum = sum + sales[i];
+            sum = sum + (int) sales[i];
         }
         return sum;
     }
 
-    public int averageSumSales(int[] sales) {
+    public int averageSumSales(long[] sales) {
         int averageSum = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            averageSum = (averageSum + sales[i]) / 2;
+            averageSum = (averageSum + (int) sales[i]) / 2;
         }
         return averageSum;
     }
 
-    public int maxSales(int[] sales) {
+    public int maxSales(long[] sales) {
         int maxMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >= sales[maxMonth]) {
+            if ((int) sales[i] >= (int) sales[maxMonth]) {
                 maxMonth = i;
             }
         }
         return maxMonth + 1;
     }
 
-    public int minSales(int[] sales) {
+    public int minSales(long[] sales) {
         int minMonth = 0;
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= sales[minMonth]) {
+            if ((int) sales[i] <= (int) sales[minMonth]) {
                 minMonth = i;
             }
         }
         return minMonth + 1;
     }
 
-    public int amountOfMinMonth(int[] sales) {
+    public int amountOfMinMonth(long[] sales) {
         int amountMinMonth = 0;
         int averageSum = averageSumSales(sales);
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] < averageSum) {
+            if ((int) sales[i] < averageSum) {
                 amountMinMonth = amountMinMonth + 1;
             }
         }
         return amountMinMonth;
     }
 
-    public int amountOfMaxMonth(int[] sales) {
+    public int amountOfMaxMonth(long[] sales) {
         int amountMaxMonth = 0;
         int averageSum = averageSumSales(sales);
 
         for (int i = 0; i < sales.length; i++) {
-            if (sales[i] > averageSum) {
+            if ((int) sales[i] > averageSum) {
                 amountMaxMonth = amountMaxMonth + 1;
             }
         }
